+++ apps/easypid/src/agent/initialize.ts
@@ -4,7 +4,7 @@ import { initializeEasyPIDAgent } from '@package/agent'
 export function initializeAppAgent({ walletKey }: { walletKey: string }) {
   return initializeEasyPIDAgent({
     keyDerivation: 'raw',
-    walletId: 'ausweis-wallet',
+    walletId: 'easypid-wallet',
     walletKey,
     walletLabel: 'EasyPID Wallet',
     trustedX509Certificates,
