+++ apps/easypid/src/constants.ts
@@ -1,5 +1,5 @@
-export const AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID = 'AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID'
-export const AUSWEIS_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID = 'AUSWEIS_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID'
+export const EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID = 'EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID'
+export const EASYPID_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID = 'EASYPID_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID'
 
 // https://demo.pid-issuer.bundesdruckerei.de
 const bdrPidIssuerCertificate = `-----BEGIN CERTIFICATE-----
