+++ apps/easypid/src/crypto/aes.ts
@@ -1,4 +1,4 @@
-import { AUSWEIS_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID } from '@easypid/constants'
+import { EASYPID_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID } from '@easypid/constants'
 import { aes256Gcm } from '@package/agent'
 
-export const ausweisAes256Gcm = aes256Gcm(AUSWEIS_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID)
+export const easyPidAes256Gcm = aes256Gcm(EASYPID_WALLET_INSTANCE_LONG_TERM_AES_KEY_ID)
