+++ apps/easypid/src/storage/seedCredential.ts
@@ -1,18 +1,18 @@
 import type { Agent } from '@credo-ts/core'
-import { AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID } from '@easypid/constants'
+import { EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID } from '@easypid/constants'
 import { walletJsonStore } from '@package/agent'
 
 export const seedCredentialStorage = {
   store: async (agent: Agent, seedCredential: string) =>
-    walletJsonStore.store(agent, AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID, {
+    walletJsonStore.store(agent, EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID, {
       seedCredential,
     }),
 
   update: async (agent: Agent, seedCredential: string) =>
-    walletJsonStore.update(agent, AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID, {
+    walletJsonStore.update(agent, EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID, {
       seedCredential,
     }),
 
   getById: async (agent: Agent) =>
-    walletJsonStore.getById<{ seedCredential: string }>(agent, AUSWEIS_WALLET_SEED_CREDENTIAL_RECORD_ID),
+    walletJsonStore.getById<{ seedCredential: string }>(agent, EASYPID_WALLET_SEED_CREDENTIAL_RECORD_ID),
 }
