+++ apps/storybook/.storybook/withThemeProvider.tsx
@@ -2,16 +2,16 @@ import { TamaguiProvider } from '@package/ui'
 import type { Decorator } from '@storybook/react'
 import React from 'react'
 
-import ausweisConfig from '../../easypid/tamagui.config'
+import easyPidConfig from '../../easypid/tamagui.config'
 import paradymConfig from '../../paradym/tamagui.config'
 
 const configs = {
-  ausweis: ausweisConfig,
+  easypid: easyPidConfig,
   paradym: paradymConfig,
 }
 
 const withThemeProvider: Decorator = (Story, context) => {
-  const configName = context.parameters.theme ?? 'ausweis'
+  const configName = context.parameters.theme ?? 'easypid'
 
   const config = configs[configName]
   if (!config)
